+++ source/main/modules/window-manager/create-ql-window.ts
@@ -17,6 +17,7 @@ import {
   BrowserWindowConstructorOptions
 } from 'electron'
 import { MDFileDescriptor } from '../fsal/types'
+import { WindowPosition } from './types'
 import setWindowChrome from './set-window-chrome'
 
 /**
@@ -26,9 +27,13 @@ import setWindowChrome from './set-window-chrome'
  * @param   {MDFileDescriptor}  file  The file to load in the Quicklook
  * @return  {BrowserWindow}           The loaded main window
  */
-export default function createQuicklookWindow (file: MDFileDescriptor): BrowserWindow {
+export default function createQuicklookWindow (file: MDFileDescriptor, conf: WindowPosition): BrowserWindow {
   const winConf: BrowserWindowConstructorOptions = {
     acceptFirstMouse: true,
+    width: conf.width,
+    height: conf.height,
+    x: conf.left,
+    y: conf.top,
     minWidth: 300,
     minHeight: 200,
     show: false,
@@ -58,6 +63,9 @@ export default function createQuicklookWindow (file: MDFileDescriptor): BrowserW
   // Only show window once it is completely initialized + maximize it
   window.once('ready-to-show', function () {
     window.show()
+    if (conf.isMaximised) {
+      window.maximize()
+    }
   })
 
   // Emitted when the user wants to close the window.